[UPDATE] TODO PROJECT 2-4 (Done)
I think the bug described in the previous commit (050b772) is caused by the fact that the three objects—local_scoreboard, lbl_list, and page_head—are not official items from the source code. That is, we have to reset the objects manually.
Although the root of this bug is currently unknown, the code runs normally now! Yay :D