[FIX] TODO PROJECT 2 File Input When File Is Empty
Originally we have a fixed line:
```
file_out << std::endl << "Anonymous " << (float) (this->GetMoney() * ((float) this->lives / (float) 100));
```
which works fine when there are already some lines in the file. However, if we delete all the content in the file, the above output command would leave a leftover `std::endl` at the beginning of the file.

In addition, when the file is empty and the user attempts to use the button `SCOREBOARD`, they would find out that the file is rewritten with a space followed by a strange number:
```
 32762
```

Now, both of the bugs are solved. 🦕