[UPDATE] TODO PROJECT 3-2 New Enemy
A new enemy type "MegaTankEnemy" has been added!
========
There is still room for improvement:
- In the Project Spec file, it reads:
```
The new turret/enemy/tool cannot be the same as, or too similar to, those in the
template.
✦ Merely changing the stats of existing turret/enemy doesn’t count.
✦ You must describe what’s special about your new turret/enemy/tool.
```
Currently my new turret and enemy highly resemble other turrets and enemies. I will resume updating the functionalities or the characteristics of the new elements later.