[UPDATE] TODO PROJECT 3 Shovel Style & MegaTankEnemy
1. Shovel Style!
Shovels need "base" images since they are of the class Turret. I replaced the "base" of the shovel with a transplant 1x1 image so that the preview of a shovel looks cleaner :)

2. MegaTankEnemy
I added color tints for the turrets when they're under the disable_radius of a MegaTankEnemy. (See MegaTankEnemy::Update() in MegaTankEnemy.cpp.) In addition, thanks to HaoDai, I discovered the issue with the original disable_cycle algorithm. I will elaborate on the current approach and my original one.